Refactor channel client-server tests.
Client/server channel tests now deal only with a pre-established
channel. This allows us to automatically create the required number of
connections per channel.
In order to guarantee the connection order on each side, the client
side send the connection id over the connection so that the server
side can match it properly.

ghstack-source-id: 099bb6af00de47b0794a3a03093b3bd1c4af3f3c
Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/288